Take fall distance as a double
Block.fallOn's fall distance widened to a double.

Error sites 1420 -> 1416.

+++ src/main/java/com/simibubi/create/content/contraptions/bearing/SailBlock.java
@@ -190,7 +190,7 @@ public class SailBlock extends WrenchableDirectionalBlock {
 	}
 
 	@Override
-	public void fallOn(Level p_152426_, BlockState p_152427_, BlockPos p_152428_, Entity p_152429_, float p_152430_) {
+	public void fallOn(Level p_152426_, BlockState p_152427_, BlockPos p_152428_, Entity p_152429_, double p_152430_) {
 		if (frame)
 			super.fallOn(p_152426_, p_152427_, p_152428_, p_152429_, p_152430_);
 		super.fallOn(p_152426_, p_152427_, p_152428_, p_152429_, 0);
